Refactor IncidentReportTracker so that the IncidentReportList can be used elsewhere in order to be able to show CMs their reports.
