+++ src/components/IncidentReportTracker/IncidentReportCard.tsx
@@ -17,7 +17,7 @@
 
 /*
  * This file contains the incident report tracking and management system
- * which is used by our IncidentReportTracker widget and our ReportsCenter view.
+ * which is used by our IncidentReportIndicator widget and our ReportsCenter view.
  */
 
 import * as React from "react";
