+++ src/components/IncidentReportTracker/IncidentReportList.styl
@@ -47,7 +47,7 @@
     }
 }
 
-.IncidentReportTracker {
+.IncidentReportList {
     .report-header {
         display: flex;
         flex-direction: row;
